Resolving MFCC results in C vs. Python Mismatch and Creating PC test for the KWS Example

+++ examples/keyword_spotting/python_speech_features/base.py
@@ -44,10 +44,13 @@ def mfcc(signal,samplerate=16000,winlen=0.025,winstep=0.01,numcep=13,
     """
     nfft = nfft or calculate_nfft(samplerate, winlen)
     feat,energy = fbank(signal,samplerate,winlen,winstep,nfilt,nfft,lowfreq,highfreq,preemph,winfunc)
+    #print(type(feat[0][0]))
+
     feat = numpy.log(feat)
-    feat = dct(feat, type=2, axis=1, norm='ortho')[:,:numcep]
+    feat = dct(feat, type=2, axis=1, norm='ortho')[:,:numcep]    
     feat = lifter(feat,ceplifter)
     if appendEnergy: feat[:,0] = numpy.log(energy) # replace first cepstral coefficient with log of frame energy
+
     return feat
 
 def fbank(signal,samplerate=16000,winlen=0.025,winstep=0.01,
@@ -70,15 +73,15 @@ def fbank(signal,samplerate=16000,winlen=0.025,winstep=0.01,
     """
     highfreq= highfreq or samplerate/2
     signal = sigproc.preemphasis(signal,preemph)
-    frames = sigproc.framesig(signal, winlen*samplerate, winstep*samplerate, winfunc)
+    frames = sigproc.framesig(signal, winlen*samplerate, winstep*samplerate, winfunc)  
     pspec = sigproc.powspec(frames,nfft)
+
+    
     energy = numpy.sum(pspec,1) # this stores the total energy in each frame
     energy = numpy.where(energy == 0,numpy.finfo(float).eps,energy) # if energy is zero, we get problems with log
-
     fb = get_filterbanks(nfilt,nfft,samplerate,lowfreq,highfreq)
     feat = numpy.dot(pspec,fb.T) # compute the filterbank energies
     feat = numpy.where(feat == 0,numpy.finfo(float).eps,feat) # if feat is zero, we get problems with log
-
     return feat,energy
 
 def logfbank(signal,samplerate=16000,winlen=0.025,winstep=0.01,
@@ -136,7 +139,7 @@ def hz2mel(hz):
     :param hz: a value in Hz. This can also be a numpy array, conversion proceeds element-wise.
     :returns: a value in Mels. If an array was passed in, an identical sized array is returned.
     """
-    return 2595 * numpy.log10(1+hz/700.)
+    return 2595 * numpy.log10(1+hz/700.).astype('float32')
 
 def mel2hz(mel):
     """Convert a value in Mels to Hertz
@@ -163,12 +166,12 @@ def get_filterbanks(nfilt=20,nfft=512,samplerate=16000,lowfreq=0,highfreq=None):
     # compute points evenly spaced in mels
     lowmel = hz2mel(lowfreq)
     highmel = hz2mel(highfreq)
-    melpoints = numpy.linspace(lowmel,highmel,nfilt+2)
+    melpoints = numpy.linspace(lowmel,highmel,nfilt+2).astype('float32')
     # our points are in Hz, but we use fft bins, so we have to convert
     #  from Hz to fft bin number
-    bin = numpy.floor((nfft+1)*mel2hz(melpoints)/samplerate)
+    bin = numpy.floor((nfft+1)*mel2hz(melpoints)/samplerate).astype('float32')
 
-    fbank = numpy.zeros([nfilt,nfft//2+1])
+    fbank = numpy.zeros([nfilt,nfft//2+1]).astype('float32')
     for j in range(0,nfilt):
         for i in range(int(bin[j]), int(bin[j+1])):
             fbank[j,i] = (i - bin[j]) / (bin[j+1]-bin[j])
